Rename component file and add index for Navigation
